requirements.txt
python-telegram-bot[httpx]>=21.0
aiosqlite>=0.20
httpx>=0.27
pydantic-settings>=2.0
nanoid>=2.0
aiohttp>=3.9
python-dotenv>=1.0
